[vm-types] Move EpochResource struct from consensus_config.rs to epoch.rs, and rename to Epoch.
