+++ chain/service/src/chain_service.rs
@@ -21,7 +21,7 @@ use starcoin_types::{
     startup_info::{ChainInfo, StartupInfo},
     transaction::{Transaction, TransactionInfo},
 };
-use starcoin_vm_types::on_chain_config::{EpochInfo, GlobalTimeOnChain};
+use starcoin_vm_types::on_chain_resource::{EpochInfo, GlobalTimeOnChain};
 use std::sync::Arc;
 
 /// A Chain reader service to provider Reader API.
