+++ cmd/starcoin/src/chain/get_global_time_by_number.rs
@@ -5,7 +5,7 @@ use crate::cli_state::CliState;
 use crate::StarcoinOpt;
 use anyhow::Result;
 use scmd::{CommandAction, ExecContext};
-use starcoin_vm_types::on_chain_config::GlobalTimeOnChain;
+use starcoin_vm_types::on_chain_resource::GlobalTimeOnChain;
 use structopt::StructOpt;
 
 /// Get global time of master.
