+++ consensus/src/cn.rs
@@ -9,7 +9,7 @@ use starcoin_crypto::HashValue;
 use starcoin_traits::ChainReader;
 use starcoin_types::block::BlockHeader;
 use starcoin_types::U256;
-use starcoin_vm_types::on_chain_config::EpochInfo;
+use starcoin_vm_types::on_chain_resource::EpochInfo;
 
 #[derive(Default)]
 pub struct CryptoNightConsensus {}
