+++ consensus/src/difficulty.rs
@@ -8,7 +8,7 @@ use anyhow::{bail, Result};
 use logger::prelude::*;
 use starcoin_traits::ChainReader;
 use starcoin_types::block::Block;
-use starcoin_vm_types::on_chain_config::EpochInfo;
+use starcoin_vm_types::on_chain_resource::EpochInfo;
 use std::convert::TryInto;
 
 /// Get the target of next pow work
