+++ core/genesis/src/lib.rs
@@ -373,7 +373,8 @@ mod tests {
     use starcoin_types::account_config::genesis_address;
     use starcoin_vm_types::account_config::association_address;
     use starcoin_vm_types::genesis_config::ChainId;
-    use starcoin_vm_types::on_chain_config::{EpochResource, VMConfig, Version};
+    use starcoin_vm_types::on_chain_config::{VMConfig, Version};
+    use starcoin_vm_types::on_chain_resource::Epoch;
 
     #[stest::test]
     pub fn test_genesis_load() -> Result<()> {
@@ -498,7 +499,7 @@ mod tests {
         block_accumulator.append(&[HashValue::random()])?;
         block_accumulator.flush()?;
 
-        let epoch = account_state_reader.get_resource::<EpochResource>(genesis_address())?;
+        let epoch = account_state_reader.get_resource::<Epoch>(genesis_address())?;
         assert!(epoch.is_some(), "Epoch resource should exist.");
 
         Ok(())
