+++ vm/types/src/on_chain_resource/mod.rs
@@ -1,7 +1,8 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod epoch;
+mod epoch;
 mod global_time;
 
+pub use epoch::{Epoch, EpochData, EpochInfo};
 pub use global_time::GlobalTimeOnChain;
